fix: taking into account GMaps API rate limit/second
GMaps API allow for 50 requests per second, which is equal to 1 request
every 20 milliseconds.
Calculating how many requests we can have in flight simultaneously: `50
* 0.2 = 10` (200ms latency is inferred, not manually calculated).
`buffer_unordered` should be at least be equal to the `semaphore` limit
or 2x greater than it.
Still unclear how to dynmically calculate a sempahore limit in that
case, latency could change and thus allow for a smaller number if less
latency if I understood correctly.